Refactor: [BADA-74] import 경로 수정 # 21
