+++ src/shared/components/ui/DatePicker/DatePicker.stories.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { DatePicker } from '@ui/DatePicker/DatePicker';
+import { DatePicker } from '@ui/DatePicker';
 
 const meta: Meta<typeof DatePicker> = {
   title: 'Components/DatePicker',
