+++ src/shared/components/ui/DatePicker/index.ts
@@ -1,0 +1,1 @@
+export * from './DatePicker';
