+++ src/shared/components/ui/LoginButton/LoginButton.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { LoginButton } from '@ui/LoginButton/LoginButton';
+import { LoginButton } from '@ui/LoginButton';
 import { useEffect } from 'react';
 import { useAuthStore } from '@features/auth/stores/authStore';
