+++ src/shared/components/ui/LoginButton/index.ts
@@ -1,0 +1,1 @@
+export { LoginButton } from './LoginButton';
