--- src/shared/components/ui/LoginButton/indext.ts
@@ -1,1 +1,0 @@
-//src/components/ui/LoginButton/index.tsxport * from './LoginButton';
